backend/requirements.txt
beautifulsoup4
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
iso8601==0.1.12
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
